requirements.txt
flask==2.3.2
flask_cors==4.0.0
flask_mail==0.9.1
PyJWT==2.8.0
boto3==1.34.65
flask_bcrypt==1.0.1
python-dotenv==1.0.1
sentence-transformers==2.2.2
scikit-learn==1.3.2
transformers==4.26.1
huggingface_hub==0.14.1
# sentence-transformers==2.2.2
requests==2.31.0
numpy==1.26.4